+++ shub/apps/library/views/images.py
@@ -13,6 +13,7 @@ from django.shortcuts import redirect, reverse
 from sregistry.utils import parse_image_name
 
 from shub.apps.logs.utils import generate_log
+from shub.apps.main.utils import format_collection_name
 from shub.apps.main.models import Collection, Container
 from shub.settings import (
     MINIO_BUCKET,
@@ -45,6 +46,7 @@ from .helpers import (
     generate_container_metadata,
     get_token,
     get_container,
+    get_collection,
     validate_token,
 )
 
@@ -532,12 +534,14 @@ class GetImageView(RatelimitMixin, APIView):
 class CollectionsView(RatelimitMixin, APIView):
     """Return a simple list of collections
     GET /v1/collections
+    Create a new collection
+    POST /v1/collections
     """
 
     ratelimit_key = "ip"
     ratelimit_rate = settings.VIEW_RATE_LIMIT
     ratelimit_block = settings.VIEW_RATE_LIMIT_BLOCK
-    ratelimit_method = "GET"
+    ratelimit_method = ("GET", "POST")
     renderer_classes = (JSONRenderer,)
 
     def get(self, request):
@@ -551,6 +555,47 @@ class CollectionsView(RatelimitMixin, APIView):
         collections = generate_collections_list(token.user)
         return Response(data=collections, status=200)
 
+    def post(self, request):
+
+        print("POST CollectionsView")
+        if not validate_token(request):
+            print("Token not valid")
+            return Response(status=403)
+
+        # body should have {'entity': entity_id, 'name': new_collection_name, 'private': bool}
+        # 'private' is optional and defaults to Collection.private default
+        # value set with get_privacy_default()
+        # {"entity": "42", "name": "my_collection", "private": true}
+        body = json.loads(request.body.decode("utf-8"))
+        if not ('entity' in body and 'name' in body):
+            print("Invalid payload")
+            return Response(status=400)
+
+        # does a collection with the same name exist already?
+        if not get_collection(body['name'], False) is None:
+            print("A collection named '{0}' exists already!".format(body['name']))
+            return Response(status=403)
+
+        # check user permission
+        token = get_token(request)
+        if str(token.user.id) != body['entity']:
+            print("Permission denied {0} {1}".format(token.user.id, body['entity']))
+            return Response(status=403)
+
+        # create the collection
+        name = format_collection_name(body['name'])
+        collection = Collection(name=name, secret=str(uuid.uuid4()))
+        collection.save()
+        collection.owners.add(token.user)
+        collection.save()
+
+        # set privacy if present
+        if 'private' in body:
+            collection.private = body['private']
+            collection.save()
+
+        details = generate_collection_metadata(collection, token.user)
+        return Response(data={"data": details}, status=200)
 
 class GetCollectionTagsView(RatelimitMixin, APIView):
     """Return a simple list of tags in the collection, and the